server/requirements.txt
Django==2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Pillow==4.3.0
